structure - Add another ItemSelector, it store the necessary tag ID in case of html, and move fieldType from FieldSelector to ItemSelector
