+++ src/main/java/com/keli/crawler/core/api/selector/item/HtmlItemSelector.java
@@ -1,0 +1,12 @@
+package com.keli.crawler.core.api.selector.item;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class HtmlItemSelector implements ItemSelector {
+
+  private Class<?> classType;
+  private String tagCssQuery;
+}
