+++ src/main/java/com/keli/crawler/core/api/selector/item/ItemSelector.java
@@ -1,0 +1,5 @@
+package com.keli.crawler.core.api.selector.item;
+
+public interface ItemSelector {
+
+}
